TRT-LLM Export Code Cleanup (#9270)
* Init code cleanup for the trt-llm export


* Removed model config


* Apply isort and black reformatting


* clearn futher


* Done more cleaning and addressing the reviews


* Apply isort and black reformatting


---------
